LFXV2-12: opensearh template query
- change the opensearch template query based on the existing poc
- transform the opensearch result to the api response
