+++ internal/service/resource_search.go
@@ -29,7 +29,7 @@ func (s *ResourceSearch) QueryResources(ctx context.Context, criteria domain.Sea
 	)
 
 	// It seems that Goa v3 does not natively support complex conditional validations
-	// like “at least one of these fields must be set
+	// like “at least one of these fields must be set"
 	if err := s.validateSearchCriteria(criteria); err != nil {
 		slog.With("error", err).ErrorContext(ctx, "search criteria validation failed")
 		return nil, fmt.Errorf("invalid search criteria: %w", err)
